requirements.txt
redis==2.10.3
python-redis-lock[django]==2.3.0